--- index.html
@@ -1,61 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <title>This Week In React</title>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta name="twitter:card" content="summary_large_image">
-  <meta name="twitter:site" content="@PhilippSpiess">
-  <meta name="twitter:title" content="This Week In React ⚛️">
-  <meta name="twitter:description" content="Weekly coverage of interesting pull requests, discussions, and RFCs around React Core and React DOM.">
-  <meta name="twitter:image" content="https://philippspiess.com/this-week-in-react/twitter-card.png">
-  <link rel="icon" type="image/png" href="./favicon.png">
-  <link rel="stylesheet" type="text/css" href="./style.css">
-  <link href="https://fonts.googleapis.com/css?family=Roboto:300,700" rel="stylesheet">
-</head>
-
-<body>
-  <div aria-hidden="true" class="atoms"></div>
-  <div class="main">
-    <section>
-      <h1>This Week In React <img src="./react.svg" width="60px" alt="⚛️" id="logo" /></h1>
-      <h2>Weekly coverage of interesting pull requests, discussions, and RFCs around React Core and React DOM. Coming out every Friday.</h2>
-      <div class="issues">
-        <ol>
-          <li><a href="https://twitter.com/PhilippSpiess/status/1050780941151588353">Issue 4 — 12 Oct 2018</a></li>
-          <li><a href="https://twitter.com/PhilippSpiess/status/1048242551998824450">Issue 3 — 5 Oct 2018</a></li>
-          <li><a href="https://twitter.com/PhilippSpiess/status/1045714338684837888">Issue 2 — 28 Sept 2018</a></li>
-          <li><a href="https://twitter.com/PhilippSpiess/status/1043186943624982529">Issue 1 — 21 Sept 2018</a></li>
-        </ol>
-      </div>
-
-      <div class="button-row">
-        <a href="https://twitter.com/PhilippSpiess">Follow on Twitter <span class="button-icon"><img src="./twitter.svg" width="30px" alt="🐦" /></span></a>
-        <a href="https://t.me/this_week_in_react">Read on Telegram <span class="button-icon"><img src="./telegram.svg" width="30px" alt="✈️" /></span></a>
-      </div>
-      <div class="button-row">
-        <a class="button" href="https://eepurl.com/dHV-3X">Subscribe Now <span class="button-icon">💌</span></a>
-      </div>
-
-      <h3>Made with 💕 by <a href="https://philippspiess.com">@PhilippSpiess</a></h3>
-    </section>
-  </div>
-
-  <script type="text/javascript" src="./atoms.js"></script>
-
-  <script type="text/javascript">
-  var clicky_site_ids = clicky_site_ids || [];
-  clicky_site_ids.push(101142968);
-  (function() {
-    var s = document.createElement('script');
-    s.type = 'text/javascript';
-    s.async = true;
-    s.src = '//static.getclicky.com/js';
-    ( document.getElementsByTagName('head')[0] || document.getElementsByTagName('body')[0] ).appendChild( s );
-  })();
-  </script>
-  <noscript><p><img alt="Clicky" width="1" height="1" src="//in.getclicky.com/101142968ns.gif" /></p></noscript>
-</body>
-
-</html>
