+++ package.json
@@ -11,11 +11,9 @@
     "format": "prettier {{src,script}/**/*.{js,css},*.md} --write"
   },
   "dependencies": {
-    "chokidar": "2.0.4",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
     "react-helmet": "^5.2.0",
-    "react-hot-loader": "^4.0.0-beta.21",
     "react-router": "^4.2.0",
     "react-router-dom": "^4.3.1",
     "react-static": "5.9.12"
